requirements.txt
discord.py
requests

# create config.py